Rename lib_machine_tests.rs to tests.rs
